isl_multi_templ.c: extract out isl_multi_arith_templ.c
These arithmetic functions only make sense for multi expressions
composed of functions, and not for isl_multi_id, which will
be introduced next.
