+++ include/isl/aff.h
@@ -449,7 +449,7 @@ __isl_give isl_set *isl_pw_aff_list_gt_set(__isl_take isl_pw_aff_list *list1,
 
 ISL_DECLARE_MULTI(aff)
 ISL_DECLARE_MULTI_CMP(aff)
-ISL_DECLARE_MULTI_NEG(aff)
+ISL_DECLARE_MULTI_ARITH(aff)
 ISL_DECLARE_MULTI_DIMS(aff)
 ISL_DECLARE_MULTI_WITH_DOMAIN(aff)
 
@@ -506,7 +506,7 @@ __isl_give isl_multi_aff *isl_multi_aff_read_from_str(isl_ctx *ctx,
 void isl_multi_aff_dump(__isl_keep isl_multi_aff *maff);
 
 ISL_DECLARE_MULTI(pw_aff)
-ISL_DECLARE_MULTI_NEG(pw_aff)
+ISL_DECLARE_MULTI_ARITH(pw_aff)
 ISL_DECLARE_MULTI_DIMS(pw_aff)
 ISL_DECLARE_MULTI_WITH_DOMAIN(pw_aff)
 
@@ -1080,7 +1080,7 @@ __isl_give isl_printer *isl_printer_print_union_pw_aff(
 void isl_union_pw_aff_dump(__isl_keep isl_union_pw_aff *upa);
 
 ISL_DECLARE_MULTI(union_pw_aff)
-ISL_DECLARE_MULTI_NEG(union_pw_aff)
+ISL_DECLARE_MULTI_ARITH(union_pw_aff)
 ISL_DECLARE_MULTI_PARAM(union_pw_aff)
 
 __isl_give isl_multi_union_pw_aff *isl_multi_union_pw_aff_from_multi_aff(
