+++ include/isl/multi.h
@@ -118,6 +118,19 @@ __isl_give isl_multi_##BASE *isl_multi_##BASE##_range_factor_domain(	\
 __isl_export								\
 __isl_give isl_multi_##BASE *isl_multi_##BASE##_range_factor_range(	\
 	__isl_take isl_multi_##BASE *multi);				\
+__isl_export								\
+__isl_give isl_multi_##BASE *isl_multi_##BASE##_align_params(		\
+	__isl_take isl_multi_##BASE *multi,				\
+	__isl_take isl_space *model);					\
+__isl_export								\
+__isl_give isl_multi_##BASE *isl_multi_##BASE##_from_range(		\
+	__isl_take isl_multi_##BASE *multi);
+
+#define ISL_DECLARE_MULTI_CMP(BASE)					\
+int isl_multi_##BASE##_plain_cmp(__isl_keep isl_multi_##BASE *multi1,	\
+	__isl_keep isl_multi_##BASE *multi2);
+
+#define ISL_DECLARE_MULTI_ARITH(BASE)					\
 __isl_overload								\
 __isl_give isl_multi_##BASE *isl_multi_##BASE##_scale_val(		\
 	__isl_take isl_multi_##BASE *multi, __isl_take isl_val *v);	\
@@ -144,19 +157,6 @@ __isl_export								\
 __isl_give isl_multi_##BASE *isl_multi_##BASE##_sub(			\
 	__isl_take isl_multi_##BASE *multi1,				\
 	__isl_take isl_multi_##BASE *multi2);				\
-__isl_export								\
-__isl_give isl_multi_##BASE *isl_multi_##BASE##_align_params(		\
-	__isl_take isl_multi_##BASE *multi,				\
-	__isl_take isl_space *model);					\
-__isl_export								\
-__isl_give isl_multi_##BASE *isl_multi_##BASE##_from_range(		\
-	__isl_take isl_multi_##BASE *multi);
-
-#define ISL_DECLARE_MULTI_CMP(BASE)					\
-int isl_multi_##BASE##_plain_cmp(__isl_keep isl_multi_##BASE *multi1,	\
-	__isl_keep isl_multi_##BASE *multi2);
-
-#define ISL_DECLARE_MULTI_NEG(BASE)					\
 __isl_export								\
 __isl_give isl_multi_##BASE *isl_multi_##BASE##_neg(		 	\
 	__isl_take isl_multi_##BASE *multi);
