+++ include/isl/val.h
@@ -13,7 +13,7 @@ extern "C" {
 #endif
 
 ISL_DECLARE_MULTI(val)
-ISL_DECLARE_MULTI_NEG(val)
+ISL_DECLARE_MULTI_ARITH(val)
 ISL_DECLARE_MULTI_DIMS(val)
 ISL_DECLARE_MULTI_WITH_DOMAIN(val)
