+++ isl_aff.c
@@ -3895,6 +3895,7 @@ error:
 #include <isl_multi_no_explicit_domain.c>
 #include <isl_multi_templ.c>
 #include <isl_multi_apply_set.c>
+#include <isl_multi_arith_templ.c>
 #include <isl_multi_cmp.c>
 #include <isl_multi_dims.c>
 #include <isl_multi_floor.c>
@@ -6411,6 +6412,7 @@ error:
 #include <isl_multi_pw_aff_explicit_domain.c>
 #include <isl_multi_templ.c>
 #include <isl_multi_apply_set.c>
+#include <isl_multi_arith_templ.c>
 #include <isl_multi_coalesce.c>
 #include <isl_multi_dims.c>
 #include <isl_multi_gist.c>
@@ -8284,6 +8286,7 @@ error:
 #include <isl_multi_templ.c>
 #include <isl_multi_apply_set.c>
 #include <isl_multi_apply_union_set.c>
+#include <isl_multi_arith_templ.c>
 #include <isl_multi_coalesce.c>
 #include <isl_multi_floor.c>
 #include <isl_multi_gist.c>
