+++ isl_val.c
@@ -1667,6 +1667,7 @@ isl_stat isl_val_check_match_domain_space(__isl_keep isl_val *v,
 #define NO_MOVE_DIMS
 #include <isl_multi_no_explicit_domain.c>
 #include <isl_multi_templ.c>
+#include <isl_multi_arith_templ.c>
 #include <isl_multi_dims.c>
 
 /* Apply "fn" to each of the elements of "mv" with as second argument "v".
